moving BuildTestError and BuildTestCommand to buildtest/utils/command.py and refactor code. Add regtest for BuildTestCommand
